init second step in game
