+++ assets/less/main.less
@@ -30,6 +30,7 @@
 @import (reference) 'project/forms';
 
 @import 'project/mainmenu';
+@import 'project/game';
 @import 'project/game-field';
 
 @import 'project/signup-form';
