+++ assets/less/project/game.less
@@ -1,0 +1,9 @@
+.game {
+  width: 100%;
+  max-width: 1000px;
+
+  &__field-container {
+    float: left;
+    margin-right: 3em;
+  }
+}
