+++ templates/game.xml
@@ -1,41 +1,17 @@
 <fest:template xmlns:fest="http://fest.mail.ru" context_name="data">
     <div class="wrapper wrapper_center">
         <h1>Игра</h1>
-        <h4><a href="#">Главное меню →</a></h4>
+        <h4><a href="#">← главное меню</a></h4>
 
-        <fest:script>data.letters = ['А', 'Б', 'В', 'Г', 'Д', 'Е', 'Ж', 'З', 'И', 'К', 'Л', 'М', 'Н', 'О'];</fest:script>
-        <fest:script>data.fields = ['left', 'right'];</fest:script>
-
-        <fest:for iterate="data.fields" index="index" value="cl">
-            <div>
-                <fest:attributes>
-                    <fest:attribute name="class">
-                        game-field-container game-field-container_<fest:value>cl</fest:value>
-                    </fest:attribute>
-                </fest:attributes>
-                <table class="game-field">
-                    <thead>
-                        <tr>
-                            <td class="game-field__cell game-field__cell_header_none"></td>
-                            <fest:for from="1" to="data.size" index="column">
-                                <td class="game-field__cell game-field__cell_header_top">
-                                    <fest:value>data.letters[column - 1]</fest:value>
-                                </td>
-                            </fest:for>
-                        </tr>
-                    </thead>
-                    <fest:for from="1" to="data.size" index="row">
-                        <tr>
-                            <td class="game-field__cell game-field__cell_header_left">
-                                <fest:value>row</fest:value>
-                            </td>
-                            <fest:for from="1" to="data.size" index="column">
-                                <td class="game-field__cell"> </td>
-                            </fest:for>
-                        </tr>
-                    </fest:for>
-                </table>
+        <div class="game">
+            <div class="game__field-container" id="field1">
+                <h3>Ваше поле</h3>
+                <fest:value output="text">data.field1</fest:value>
             </div>
-        </fest:for>
+            <div class="game__field-container" id="field2">
+                <h3>Поле соперника</h3>
+                <fest:value output="text">data.field2</fest:value>
+            </div>
+        </div>
     </div>
 </fest:template>
